\name{IRTtestinfo-package}

\alias{IRTtestinfo-package}

\docType{package}

\title{IRTtestinfo}

\description{This package provides functions that estimate the levels of attenuation in 
	correlation that occurs for two variables, given the IRT test information values for
	the two measures. The focus is on the effect size attenuation that is caused by the
	(often fluctuating) levels of measurement error. It also provides the test information
	functions for a number of assessment instruments in the literature.
	}

