\name{zscores_data_PCL_SV_NEO_FFI_A}
\alias{zscores_data_PCL_SV_NEO_FFI_A}
\docType{data}
\title{zscores_data_PCL_SV_NEO_FFI_A}
\description{A data frame with respondent z scores on the NEO-FFI Antagonism-Agreeableness 
	scale and on the PCL-SV scale, based on responses from 1229 participants 
	in the MacArthur Violence Risk Assessment Study dataset (Monahan et al., 2001).}
\usage{data(zscores_data_PCL_SV_NEO_FFI_A)}
\source{Monahan, J., Steadman, H., Silver, E., Appelbaum, P., Robbins, P., Mulvey, E., 
	Roth, L., Grisso, T., & Banks, S. (2001). Rethinking Risk Assessment: The MacArthur 
	Study of Mental Disorder and Violence. New York: Oxford University Press.}
\examples{
\dontrun{
head(zscores_data_PCL_SV_NEO_FFI_A)
     
}
}
